Run refill on faucet start
